Modify ChoiceTemplate reopsitory and interactor
